routing in the uploader
